- Cleanup files: Move RefitService to Services folder and move unused files to deprecated folder
