+++ src/Xablu.WebApiClient/Attributes/NameOfItemAttribute.shared.cs
