+++ src/Xablu.WebApiClient/Client/DeprecatedForNow/IRefitApi.cs
