+++ src/Xablu.WebApiClient/Client/DeprecatedForNow/RefitApi.cs
