+++ src/Xablu.WebApiClient/Client/DeprecatedForNow/RefitClient.cs
